backend/requirements.txt
Django~=5.1.6
djangorestframework~=3.15.2
django-rest-knox~=5.0.2
django-rest-passwordreset~=1.5.0
psycopg2-binary==2.9.9

python-dotenv>=1.0.0
pymupdf>=1.22.5
sentence-transformers>=2.2.2
opencv-python>=4.5.5
pytesseract>=0.3.10
numpy>=1.21.6
scikit-learn>=1.0.2
pandas>=1.3.5

django-filter>=23.2
xlsxwriter>=3.1.0
python-dateutil>=2.8.2


reportlab>=4.0.4
pillow>=10.0.0

django-celery-beat>=2.5.0

channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

transformers>=4.35.0
torch~=2.6.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
redis~=5.2.1
pdf2image~=1.16.3
requests~=2.32.3
huggingface-hub~=0.30.2
celery~=5.5.2
llama-cpp-python>=0.2.0
psutil>=5.9.0

